♻️ Refactor : 태그 삭제 조건 추가 (#34)
* refactor: doneCount 추가

member 엔티티에 doneCount 추가

* fix : 엔티티 필드 수정

- todo 엔티티 내 예약어 사용되어 수정
   - year, month, day => todoYear, todoMonth, todoDay
- tag 엔티티 필드명 수정
  - 식별을 위해 tag 단어 붙임 => tagName, tagColor

* feat : todo등록 구현(시큐리티 구현 필요)

* feat: 태그 등록 수정 삭제 구현

태그 등록 수정 삭제 기능 추가

* feat : 예외구현

- CustomErrorCodeEnum 활용하여 커스텀 가능
- API 명세대로 구현

* feat : 예외구현

- CustomErrorCodeEnum 활용하여 커스텀 가능
- API 명세대로 구현

* feat : 투두 수정 구현

* fix : Timestamp 나노초 제거 반영

* fix : 오타수정

* refactor: CustomException 적용, 공통된 메서드 수정

태그 수정, 삭제 부분에서 예외처리를 CustomException으로 변경 및 공통된 예외처리를 하나의 메서드로 분리해서 처리

* feat : 투두 삭제 구현

- 로그인 기능 구현 완료 시, 관련 내용 반영 필요

* fix : todo 및 tag 예외 enum 수정

- _MSG 문자열 추가

* refactor : todo 예외 enum 수정 및 todo service 관련 내용 반영

- _MSG 문자열 추가

* refactor : Todo 주석 수정

- 기능 명칭 부분 // -> /** **/ 로 변경

* feat : 전체조회(월단위) 구현

- 스프링시큐리티 반영 필요

* refactor: Security 적용

전반적인 태그 로직에 Security 적용

* feat: 태그 조회 기능

Security를 사용해 로그인한 사용자가 태그를 조회 할 수 있도록 구현

* refactor: 태그 삭제 조건 추가

태그가 DB에 한개라도 있으면 삭제가 불가능하도록 수정

* fix: 태그 Controller, Service에 List Import 추가

태그 Controller, Service에 List Import 추가

+++ src/main/java/com/team6/todomateclone/tag/service/TagService.java
@@ -52,7 +52,7 @@ public class TagService {
     // 태그 수정
     @Transactional
     public ResponseTagDto updateTag(Long tagId, RequestTagDto requestTagDto, Member member) {
-        // 태그 존재 여부 확인
+        // 선택한 태그 존재 여부 확인
         Tag tag = checkTag(tagId);
         // 태그 유효성 검사
         checkPermission(member, tag);
@@ -65,12 +65,17 @@ public class TagService {
     // 태그 삭제
     @Transactional
     public void deleteTag(Long tagId, Member member) {
-        // 태그 존재 여부 확인
+        // 선택한 태그 존재 여부 확인
         Tag tag = checkTag(tagId);
         // 태그 유효성 검사
         checkPermission(member, tag);
-
-        tagRepository.deleteById(tagId);
+        // 태그가 하나라도 있으면 삭제 불가
+        List<Tag> tagExist = tagRepository.findAll();
+        if (tagExist.size()>1) {
+            tagRepository.deleteById(tagId);
+        } else {
+            throw new CustomErrorException(CustomErrorCodeEnum.TAG_NOT_DELETE_MSG);
+        }
     }
 
     // 태그가 있는지 없는지 확인하는 메서드
